requirements.txt
pyquery
javaobj-py3
pycryptodome
pillow
requests
# pysqlcipher3==1.2.0
rotki-pysqlcipher3==2024.10.1
csscompressor
numpy
ipython
websocket-client
